requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==41.0.7
marshmallow==3.20.1
flasgger==0.9.7.1
bcrypt==4.1.2
